+++ src/hooks/useStatistics/index.tsx
@@ -12,26 +12,18 @@ import {
   MoodPeaksNegativeData,
   MoodPeaksPositiveData
 } from "./MoodPeaks";
-import {
-  defaultMoodTrendData,
-  getMoodTrendData,
-  MoodTrendData
-} from "./MoodTrend";
 import { defaultStreaksData, getCurrentStreak, getLongestStreak, StreaksData } from "./Streaks";
 import {
   defaultTagsDistributionData,
   getTagsDistributionData,
   TagsDistributionData
 } from "./TagsDistribution";
-import {
-  defaultTagsDistributionTrendData,
-  getTagsDistributionTrendData,
-  TagsDistributionTrendData
-} from "./TagsDistributionTrend";
 import { getTagsPeaksData, TagsPeakData } from "./TagsPeaks";
+import { EmotionsDistributionData, defaultEmotionsDistributionData, getEmotionsDistributionData } from "./EmotionsDistributuon";
+import { SleepQualityDistributionData, defaultSleepQualityDistributionDataForXDays, getSleepQualityDistributionForXDays } from "./SleepQualityDistribution";
+import { DATE_FORMAT } from "@/constants/Config";
 
 const DELAY_LOADING = 1 * 1000;
-
 
 export const STATISTIC_TYPES = [
   "mood_avg",
@@ -49,18 +41,17 @@ interface StatisticsState {
   moodAvgData: MoodAvgData;
   moodPeaksPositiveData: MoodPeaksPositiveData;
   moodPeaksNegativeData: MoodPeaksNegativeData;
+  emotionsDistributionData: EmotionsDistributionData;
   tagsPeaksData: TagsPeakData;
   tagsDistributionData: TagsDistributionData;
+  sleepQualityDistributionData: SleepQualityDistributionData;
   streaks: StreaksData;
-  trends: {
-    moodData: MoodTrendData;
-    // tagsDistributionData: TagsDistributionTrendData;
-  };
 }
 
 interface Value {
   load: ({ force }: { force: boolean }) => void;
   isAvailable: (type: StatisticType) => boolean;
+  isHighlighted: (type: StatisticType) => boolean;
   isLoading: boolean;
   state: StatisticsState;
 }
@@ -84,14 +75,12 @@ export function StatisticsProvider({
     moodAvgData: defaultMoodAvgData,
     moodPeaksPositiveData: defaultMoodPeaksPositiveData,
     moodPeaksNegativeData: defaultMoodPeaksNegativeData,
+    emotionsDistributionData: defaultEmotionsDistributionData,
     tagsDistributionData: defaultTagsDistributionData,
+    sleepQualityDistributionData: defaultSleepQualityDistributionDataForXDays(),
     streaks: defaultStreaksData,
     tagsPeaksData: {
       tags: [],
-    },
-    trends: {
-      moodData: defaultMoodTrendData,
-      // tagsDistributionData: defaultTagsDistributionTrendData,
     },
   });
 
@@ -122,11 +111,9 @@ export function StatisticsProvider({
       tags
     );
 
-    const moodTrendData = getMoodTrendData(trendsItems);
-    // const tagsDistributionTrendData = getTagsDistributionTrendData(
-    //   trendsItems,
-    //   tags
-    // );
+    const emotionsDistributionData = getEmotionsDistributionData(highlightItems);
+
+    const sleepQualityDistributionData = getSleepQualityDistributionForXDays(highlightItems, dayjs().subtract(14, "day").format(DATE_FORMAT), 30);
 
     const newState = {
       loaded: true,
@@ -136,13 +123,11 @@ export function StatisticsProvider({
       moodPeaksNegativeData,
       tagsPeaksData,
       tagsDistributionData,
+      emotionsDistributionData,
+      sleepQualityDistributionData,
       streaks: {
         longest: getLongestStreak(logState.items),
         current: getCurrentStreak(logState.items),
-      },
-      trends: {
-        moodData: moodTrendData,
-        // tagsDistributionData: tagsDistributionTrendData,
       },
     };
 
@@ -173,22 +158,64 @@ export function StatisticsProvider({
     if (type === "tags_distribution") {
       return state.tagsDistributionData?.tags.length > 0;
     }
-    // if (type === "tags_distribution_trend") {
-    //   return state.trends.tagsDistributionData.tags.length > 0;
-    // }
-    if (type === "mood_trend") {
-      return (
-        state.trends.moodData.avgPeriod1 > 0 &&
-        state.trends.moodData.avgPeriod2 > 0 &&
-        state.trends.moodData.diff > 0
-      )
+    if (type === "emotions_distribution") {
+      return state.emotionsDistributionData?.emotions.length > 3;
+    }
+    if (type === "sleep_quality_distribution") {
+      return state.sleepQualityDistributionData?.length > 7;
     }
     return false;
   };
 
+  const isHighlighted = (type: typeof STATISTIC_TYPES[number]) => {
+    if (type === "mood_avg") {
+      return (
+        isAvailable(type) &&
+        state.moodAvgData.ratingHighestPercentage > 60
+      );
+    }
+
+    if (type === "mood_peaks_positive") {
+      return (
+        isAvailable(type) &&
+        state.moodPeaksPositiveData.days.length >= 2
+      );
+    }
+
+    if (type === "mood_peaks_negative") {
+      return (
+        isAvailable(type) &&
+        state.moodPeaksNegativeData.days.length >= 2
+      );
+    }
+
+    if (type === "tags_peaks") {
+      return (
+        isAvailable(type) &&
+        state.tagsPeaksData.tags.filter((tag) => tag.items.length > 5).length > 0
+      );
+    }
+
+    if (type === "tags_distribution") {
+      return (
+        isAvailable(type)
+      )
+    }
+
+    if (type === "emotions_distribution") {
+      return (
+        isAvailable(type) &&
+        state.emotionsDistributionData.emotions.some((emotion) => emotion.count > 5)
+      );
+    }
+
+    return false;
+  }
+
   const value: Value = {
     load,
     isAvailable,
+    isHighlighted,
     isLoading,
     state,
   };