+++ src/hooks/useTemporaryLog.tsx
@@ -11,6 +11,7 @@ export type TemporaryLogState = Omit<LogItem, 'rating' | 'sleep'> & {
 export interface TemporaryLogValue {
   data: TemporaryLogState;
   isDirty: boolean;
+  isInitialized: boolean;
   initialize: (log: TemporaryLogState) => void;
   set: (log: TemporaryLogState) => void;
   update: (log: Partial<TemporaryLogState>) => void;
@@ -22,6 +23,7 @@ const TemporaryLogStateContext = createContext({} as TemporaryLogValue);
 function TemporaryLogProvider({ children }: { children: React.ReactNode }) {
   const [isDirty, setIsDirty] = useState(false);
   const [temporaryLog, setTemporaryLog] = useState<TemporaryLogState>({} as TemporaryLogState);
+  const [isInitialized, setIsInitialized] = useState(false);
 
   const update = (next: Partial<TemporaryLogState>) => {
     setTemporaryLog((current) => {
@@ -35,6 +37,7 @@ function TemporaryLogProvider({ children }: { children: React.ReactNode }) {
 
   const initialize = (log: TemporaryLogState) => {
     setTemporaryLog(log);
+    setIsInitialized(true);
   };
 
   const set = (log: TemporaryLogState) => {
@@ -45,6 +48,7 @@ function TemporaryLogProvider({ children }: { children: React.ReactNode }) {
   const reset = () => {
     setTemporaryLog({} as TemporaryLogState);
     setIsDirty(false);
+    setIsInitialized(false);
   };
 
   const value = useMemo(() => ({
@@ -57,7 +61,15 @@ function TemporaryLogProvider({ children }: { children: React.ReactNode }) {
   }), [temporaryLog, isDirty]);
 
   return (
-    <TemporaryLogStateContext.Provider value={value}>
+    <TemporaryLogStateContext.Provider value={{
+      data: temporaryLog,
+      initialize,
+      set,
+      update,
+      reset,
+      isDirty,
+      isInitialized,
+    }}>
       {children}
     </TemporaryLogStateContext.Provider>
   );
@@ -73,7 +85,7 @@ function useTemporaryLog(defaultValue?: TemporaryLogState): TemporaryLogValue {
   }
 
   useEffect(() => {
-    if (defaultValue) {
+    if (defaultValue && !context.isInitialized) {
       context.initialize(defaultValue);
     }
   }, []);