+++ src/screens/Calendar/CalendarFooter.tsx
@@ -1,3 +1,4 @@
+import * as Updates from 'expo-updates';
 import Button from "@/components/Button";
 import { t } from "@/helpers/translation";
 import useColors from "@/hooks/useColors";
@@ -8,6 +9,8 @@ import React from "react";
 import { Text, View } from "react-native";
 import { PlusCircle } from "react-native-feather";
 import { PromoCards } from "./PromoCards";
+import { FeedbackBox } from "../LogList/FeedbackBox";
+import { Armchair } from "lucide-react-native";
 
 export const CalendarFooter = () => {
   const colors = useColors();
@@ -27,26 +30,59 @@ export const CalendarFooter = () => {
           marginTop: 24,
         }}
       >
-        {!hasTodayItem ? (
-          <Button
-            icon={<PlusCircle width={24} height={24} color={colors.primaryButtonText} />}
-            onPress={() => {
-              navigation.navigate("LogCreate", {
-                dateTime: dayjs().toISOString(),
-              });
-            }}
-          >{t('add_today_entry')}</Button>
-        ) : (
-          <Button
-            icon={<PlusCircle width={24} height={24} color={colors.tertiaryButtonText} />}
-            type="tertiary"
-            onPress={() => {
-              navigation.navigate("LogCreate", {
-                dateTime: dayjs().toISOString(),
-              });
-            }}
-          >{t('add_today_another_entry')}</Button>
-        )}
+        <View
+          style={{
+          }}
+        >
+          {!hasTodayItem ? (
+            <Button
+              icon={<PlusCircle width={24} height={24} color={colors.primaryButtonText} />}
+              onPress={() => {
+                navigation.navigate("LogCreate", {
+                  dateTime: dayjs().toISOString(),
+                });
+              }}
+            >{t('add_today_entry')}</Button>
+          ) : (
+            <Button
+              icon={<PlusCircle width={24} height={24} color={colors.tertiaryButtonText} />}
+              type="tertiary"
+              onPress={() => {
+                navigation.navigate("LogCreate", {
+                  dateTime: dayjs().toISOString(),
+                });
+              }}
+            >{t('add_today_another_entry')}</Button>
+          )}
+
+          {Updates.channel !== 'production' && (
+            <>
+              <Button
+                type="secondary"
+                icon={<Armchair size={24} color={colors.secondaryButtonText} />}
+                onPress={() => {
+                  navigation.navigate("LogCreate", {
+                    dateTime: dayjs().toISOString(),
+                    avaliableSteps: ['rating'],
+                  });
+                }}
+                style={{
+                  flex: 1,
+                  marginTop: 8,
+                }}
+              >{t('quick_checkin')}</Button>
+
+              <FeedbackBox
+                emoji="📌"
+                prefix="beta_different_checkins"
+                style={{
+                  marginTop: 16,
+                  marginBottom: 0,
+                }}
+              />
+            </>
+          )}
+        </View>
       </View>
       <PromoCards />
     </View>