+++ src/screens/DevelopmentTools.tsx
@@ -12,6 +12,7 @@ import useColors from "../hooks/useColors";
 import { useLogState } from "../hooks/useLogs";
 import { useSettings } from "../hooks/useSettings";
 import { useTagsState } from "../hooks/useTags";
+import { Trash } from "lucide-react-native";
 
 const Card = ({
   title,
@@ -164,7 +165,11 @@ export const DevelopmentTools = () => {
                   justifyContent: "space-between",
                 }}
               >
-                <View>
+                <View
+                  style={{
+                    maxWidth: "85%",
+                  }}
+                >
                   <Text
                     style={{
                       color: colors.text,
@@ -185,11 +190,10 @@ export const DevelopmentTools = () => {
                   </Text>
                 </View>
                 <LinkButton
-                  icon="Trash"
                   onPress={() => {
                     removeActionDone(action.title);
                   }}
-                />
+                ><Trash size={20} color={colors.tint} /></LinkButton>
               </View>
             </MenuListItem>
           ))}
@@ -227,29 +231,6 @@ export const DevelopmentTools = () => {
         <TextInfo>
           {t("development_statistics_reset_questions_description")}
         </TextInfo>
-        <MenuListHeadline>Tags</MenuListHeadline>
-        <MenuList style={{}}>
-          {tags.map((tag, i) => (
-            <MenuListItem
-              style={{
-                flexDirection: "column",
-              }}
-              key={i}
-              title={tag.title}
-              isLast={i === tags.length - 1}
-            >
-              <Text
-                style={{
-                  color: colors.textSecondary,
-                  fontSize: 14,
-                  marginTop: 4,
-                }}
-              >
-                {tag.id}
-              </Text>
-            </MenuListItem>
-          ))}
-        </MenuList>
         <View
           style={{
             height: 100,