+++ src/screens/LogList/Sleep.tsx
@@ -1,9 +1,9 @@
 import { SlideSleepButton } from '@/components/Logger/slides/SlideSleepButton';
+import { LogItem } from '@/hooks/useLogs';
+import { useNavigation } from '@react-navigation/native';
 import { t } from 'i18n-js';
 import { View } from 'react-native';
-import { LogItem } from '@/hooks/useLogs';
-import { Headline } from './Headline';
-import { useNavigation } from '@react-navigation/native';
+import { SectionHeader } from './SectionHeader';
 
 export const Sleep = ({
   item,
@@ -17,10 +17,17 @@ export const Sleep = ({
   return (
     <View
       style={{
-        marginTop: 24,
       }}
     >
-      <Headline>{t('view_log_sleep')}</Headline>
+      <SectionHeader
+        title={t('view_log_sleep')}
+        onEdit={() => {
+          navigation.navigate('LogEdit', {
+            id: item.id,
+            step: 'sleep',
+          });
+        }}
+      />
       <View
         style={{
           flexDirection: 'row',