+++ src/screens/index.tsx
@@ -13,3 +13,4 @@ export * from './TagCreate';
 export * from './Statistics/Highlights';
 export * from './BotLogger';
 export * from './LogList'
+export * from './Settings/Tags'